FIX: Elevation Angle for some products
Many Nexrad Level 3 files have elevation angle defined by Halfword 30 of the product description block, but not all. (Elevation angle is not relevant for all products.)

Added a list of products for which elevation angle is defined by Halfword 30 per Table V of the ICD. For products for which it is not defined added logic to return data as None.